requirements.txt
dkg>=8.0.11
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.1
networkx>=3.2.1
numpy>=1.26.2
requests>=2.31.0
web3>=6.11.3
